refactor(dashboard): extract milestone celebration lifecycle into hook
Move the localStorage-backed seen tracking, unlock/tier-complete toasts,
and transient celebrating-set state (4 effects + refs + 2 keys) verbatim
into _hooks/use-milestone-celebrations.ts. Owns its own
useTranslations("Dashboard.body"); returns only what the JSX consumes
(celebratingMilestones, celebratingTiers, openMilestoneId,
setOpenMilestoneId). Behavior-identical; page.tsx 1867->1706 LOC.
